Build and release AMIs with GitHub Actions
The CI now builds the VHDs and uploads them to S3.
A targeted Terraform plan picks up the AMIs and then a full plan creates
the EBS snapshots, AMIs, permissions, etc. The source AMI is copied
to the other regions within Terraform.
Lastly, the new AMI ids are merged with the existing ids and a PR is
opened.